# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ServerAPi for Rokhelper.",
        "title": "Fiber Swagger Example API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/config/{id}": {
            "get": {
                "description": "Get the title configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit title configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Edit config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/maps/{id}": {
            "put": {
                "description": "Edit home and lost kingdom maps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "map"
                ],
                "summary": "Edit maps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Maps",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/title/done/{id}": {
            "post": {
                "description": "Mark a title assignment as done",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "title"
                ],
                "summary": "Done title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Title Assignment",
                        "name": "titleAssignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/title/finish/{id}": {
            "post": {
                "description": "Mark a title assignment as finished",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "title"
                ],
                "summary": "Finish title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Title Assignment",
                        "name": "titleAssignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/title/{id}": {
            "get": {
                "description": "Retrieve all titles for a Kingdom or Discord Channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "title"
                ],
                "summary": "Get titles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new title assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "title"
                ],
                "summary": "Add title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kingdom ID or Discord Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Title Assignment",
                        "name": "titleAssignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TitleAssignment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Config": {
            "type": "object",
            "properties": {
                "architect": {
                    "description": "Thời gian giữ title Kiến trúc sư",
                    "type": "integer"
                },
                "duke": {
                    "description": "Thời gian giữ title Công tước",
                    "type": "integer"
                },
                "justice": {
                    "description": "Thời gian giữ title Công lý",
                    "type": "integer"
                },
                "scientist": {
                    "description": "Thời gian giữ title Nhà khoa học",
                    "type": "integer"
                }
            }
        },
        "model.Local": {
            "type": "object",
            "properties": {
                "map": {
                    "description": "map chỉ có thể có 2 dữ liệu là \"home\" hoặc \"lost\"",
                    "type": "string"
                },
                "x": {
                    "type": "string"
                },
                "y": {
                    "type": "string"
                }
            }
        },
        "model.TitleAssignment": {
            "type": "object",
            "properties": {
                "local": {
                    "description": "Vị trí của người chơi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Local"
                        }
                    ]
                },
                "player_id": {
                    "description": "ID của người chơi",
                    "type": "string"
                },
                "time_add": {
                    "description": "Thời gian thêm title vào hàng đợi",
                    "type": "integer"
                },
                "title": {
                    "description": "Tên của title (Duke, Architect, Scientist, Justice)",
                    "type": "string"
                }
            }
        }
    }
}